fix del spinner vacio

+++ app/src/main/java/com/botigocontigo/alfred/areas/AreasFragment.kt
@@ -59,8 +59,7 @@ class AreasFragment : Fragment(), View.OnClickListener{
         val services = Services(inflater.context)
         api = services.botigocontigoApi()
 
-        if (executeInitializeDB)
-            api.areasGetAll().call(AreasGetCallbacks(::loadToDB))
+        api.areasGetAll().call(AreasGetCallbacks(::loadToDB))
 
         return vfrag
     }
@@ -87,8 +86,8 @@ class AreasFragment : Fragment(), View.OnClickListener{
 
     private fun switchFragment(areaName: String){
         //Cambio al Fragment de Detalle de area
-
         executeInitializeDB = false
+        Log.i("areas en el 1ro", model.toString())
 
         fragmentManager!!
                 .beginTransaction()
@@ -138,6 +137,15 @@ class AreasFragment : Fragment(), View.OnClickListener{
         val spinner = vfrag?.findViewById<Spinner>(R.id.spinner_modelos)
         spinner!!.adapter = ArrayAdapter(context, android.R.layout.simple_spinner_item, mapModels.keys.toList()).apply {
             setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
+        }
+        spinner.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
+            override fun onNothingSelected(parent: AdapterView<*>?) { }
+            override fun onItemSelected(parent: AdapterView<*>?, view2: View?, position: Int, id: Long) {
+                val modelName = spinner_modelos.selectedItem.toString()
+                doAsync {
+                    model = areaDao.findById(mapModels.getValue(modelName)!!)
+                }
+            }
         }
 
         Log.i("MAP ITEMS", mapModels.toList().toString())
@@ -207,36 +215,29 @@ class AreasFragment : Fragment(), View.OnClickListener{
     }
 
     private fun loadToDB(areas: List<Area>) {
+
         doAsync {
-            areaDao.deleteAllRows()
-            if (areas.isNotEmpty()) {
-                areaDao.insertAll(*areas.toTypedArray())
-            } else {
-                areaDao.insertAll(
-                        Area("1", "1", "Modelo A", "YPF, Repsol, AXION","relA","chanC","valueA","actA","resoA","parA","incA","costA"),
-                        Area("2", "1", "Modelo B", "clientesB","relB","chanB","valueB","actB","resoB","partB","incB","costB"),
-                        Area("3", "1", "Modelo C", "clientesC","relC","chanC","valueC","actC","resoC","partC","incC","costC")
-                )
+
+            if (executeInitializeDB) {
+                areaDao.deleteAllRows()
+                if (areas.isNotEmpty()) {
+                    areaDao.insertAll(*areas.toTypedArray())
+                } else {
+                    areaDao.insertAll(
+                            Area("1", "1", "Modelo A", "YPF, Repsol, AXION","relA","chanC","valueA","actA","resoA","parA","incA","costA"),
+                            Area("2", "1", "Modelo B", "clientesB","relB","chanB","valueB","actB","resoB","partB","incB","costB"),
+                            Area("3", "1", "Modelo C", "clientesC","relC","chanC","valueC","actC","resoC","partC","incC","costC")
+                    )
+                }
             }
 
             //HARDCODED USER ID
             mapModels = areaDao.getAll().map { it.name to it.id }.toMap()
 
-            spinner_modelos.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
-                override fun onNothingSelected(parent: AdapterView<*>?) { }
-                override fun onItemSelected(parent: AdapterView<*>?, view2: View?, position: Int, id: Long) {
-                    val modelName = spinner_modelos.selectedItem.toString()
-                    doAsync {
-                        model = areaDao.findById(mapModels.getValue(modelName)!!)
-                    }
-                }
-            }
-
-
-
-            Log.i("DAOO CREATE: ", areaDao.getAll().toString())
-
             uiThread {
+
+                Log.i("mapModels",mapModels.toString())
+
                 loadEventOnClickNewModel()
                 loadSpinnerModelos()
             }
